+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make bar chart a subcomponent of cartesian chart.
 - Make grouped bar chart a subcomponent of cartesian chart.
 - Make stacked bar chart a subcomponent of cartesian chart.
+- Make point chart a subcomponent of cartesian chart.
 
 ### Fixed